requirements.txt
langchain>=0.3.1,<0.4.0
openai>=1.58.1,<2.0.0
chromadb>=0.5.10,<0.6.0
python-dotenv>=1.0.0
tiktoken>=0.7.0 